Move ApiServerConfig
