+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/auth/JaasPrincipal.java
@@ -17,11 +17,14 @@ package io.confluent.ksql.api.auth;
 
 import java.security.Principal;
 
-class ApiPrincipal implements Principal {
+/**
+ * Principal implementation created when authenticating with the JaasAuthProvider
+ */
+class JaasPrincipal implements Principal {
 
   private final String name;
 
-  ApiPrincipal(final String name) {
+  JaasPrincipal(final String name) {
     this.name = name;
   }
