+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/api/ListenersTest.java
@@ -19,7 +19,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 
-import io.confluent.ksql.api.auth.ApiServerConfig;
+import io.confluent.ksql.api.server.ApiServerConfig;
 import io.confluent.ksql.test.util.secure.ServerKeyStore;
 import java.net.URI;
 import java.util.HashMap;
