+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/api/MaxQueriesTest.java
@@ -21,7 +21,7 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
-import io.confluent.ksql.api.auth.ApiServerConfig;
+import io.confluent.ksql.api.server.ApiServerConfig;
 import io.confluent.ksql.api.server.PushQueryId;
 import io.confluent.ksql.api.utils.QueryResponse;
 import io.vertx.core.buffer.Buffer;
